AT_04.18.02 | Label selected trip card
