+++ cypress/pageObjects/CreateBookingPage.js
@@ -37,13 +37,14 @@ class CreateBookingPage {
     getLabelDepartureOnDate = () => cy.get('#label-departure-on #trips-selected-date');
     getNumberAllSeatsFirstTripCard = () => cy.get('div .trip:first-child .class');
     getPriceOfTicket = () => cy.get('.vehclass-ferry.selected .price');
-    getLabelTicketsAvailableSelectedTripCard = () => cy.get('.selected .availability');
+    getLabelTicketsAvailableSecondTripCard = () => cy.get('.trip:nth-child(2) span.availability');
     getDepartureTime = () => cy.get('.popup-trip div:nth-child(6) span');
     getDepartureLatestButton = () => cy.get('button.trips-order-desc');
     getDepartureTripCardsList = () => cy.get('.trips-list-wrapper > div.trip');
     getBtnErliest = () => cy.get('button.trips-order-asc');
     getTripClassDropdown = () => cy.get('select[name="trips_class"]');
     getVehicleClassTripCards = () => cy.get('.trips-list-wrapper > div> span.class');
+    getNumberTicketsAvailableSecondTripCard = () => cy.get('.trip:nth-child(2) span.availability .num');
 
     //Arrival on
     getArrivalTime = () => cy.get('.popup-trip div:nth-child(7) span');
@@ -391,5 +392,7 @@ class CreateBookingPage {
         this.getMainPassengerFareTypeDropdownSelect()
             .select(FareType, { force: true })        
     }
+
+    getLableNumberAvailable
 }
 export default CreateBookingPage;
